Inline idl.js
According to the git history this was originally extracted to a
separate file to centralize this logic (i.e. simplify
maintainability), but this hasn't been a concern for years since I use
Pug nowadays. So we inline to improve performance by shaving a
request/response cycle.

It's at the end of the <body> which means that it won't block load,
but it also means that it won't load as fast. I will fix this shortly
with a rel=preload directive.

Resolves #83