Fix broken blog page links (for real)
The problem was this missing trailing slash, which caused relative links
including `..` components to resolve to incorrect
pages. I.e. `../[number]/` on `/blog/page/[number2]` would incorrectly
resolve to `/blog/[number2]` (whereas `/blog/page/[number2]/` would
resolve correctly, because in the former the number is considered a
page, whereas in the latter it's considered a directory).

So why hasn't this problem cropped up before? Well, until recently,
Apache would end up redirecting this to the proper location server-side
because the client was requesting what was a directory on disk even
though the request looked like it was for a page. So even though there
wasn't a trailing `/` in the HTML link's URL, it still ended up in the
right place.

But recently I landed Turbolinks support (see #124). Turbolinks fixes
single-page navigation not working for non-navbar links, including these
page links. But it also means that server-side redirects aren't
reflected properly in the URL bar, and that was what uncovered this bug
(which apparently has been lying latent literally since the day I
enabled pagination support at all).

The solution, of course, is simply to properly append `/` to the URLs in
the HTML source. That way we don't have to rely on server-side
redirects, which is a performance improvement anyway, and `..` works
properly now too.